Replace Failure with Mssql_error
This is a structure error for any Mssql error, which lets us attach
the source code position, query, params, formatted query, reraised
exn, etc.

The main benefit of this is that the error message will be at the
top, so it should be easier to see at-a-glance what happened with
a Sentry error.