Switch from Dblib.canquery to Dblib.cancel
See docs:

http://infocenter.sybase.com/help/index.jsp?topic=/com.sybase.help.ocs_12.5.1.dblib/html/dblib/X15355.htm
http://infocenter.sybase.com/help/index.jsp?topic=/com.sybase.help.ocs_12.5.1.dblib/html/dblib/X57019.htm

dbcanquery cancels the current query and discards results.

dbcancel cancels all current commands discards results.

We're calling this to ensure the connection is in a good state, so
dbcancel is what we want.

This should fix occasional errors where we get an exception
because a command in still in progress (presumably an incomplete
command from an exception).